Move boundary::balancer to infra::balancer::vault
