+++ crates/contracts/src/lib.rs
@@ -1,7 +1,11 @@
 #![allow(clippy::let_unit_value)]
+
+pub use ethcontract;
+
 #[cfg(feature = "bin")]
 pub mod paths;
 pub mod vault;
+pub mod web3;
 
 macro_rules! include_contracts {
     ($($name:ident;)*) => {$(
