+++ crates/solvers/src/boundary/mod.rs
@@ -1,7 +1,6 @@
 //! This is a abstraction layer between the solver engine and the existing
 //! "legacy" logic in `shared` and `model`.
 
-pub mod balancer;
 pub mod baseline;
 pub mod legacy;
 pub mod liquidity;
